Moved SpotifyConnectionInterface to UseCase SearchSong
