+++ src/main/java/use_case/playlist_user_story/search_song/SpotifyConnectionInterface.java
@@ -1,4 +1,4 @@
-package spotify_api;
+package use_case.playlist_user_story.search_song;
 
 import se.michaelthelin.spotify.model_objects.specification.Track;
